apply fixes to dialogue as well
